finish first draft of new design
