+++ package.json
@@ -9,6 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@metamask/jazzicon": "^2.0.0",
     "@react-spring/parallax": "^9.7.2",
     "@tippyjs/react": "^4.2.6",
