+++ src/components/patient-record/index.ts
@@ -1,0 +1,1 @@
+export { PatientRecord } from "./patient-record";
