+++ src/components/patient-record/patient-record.module.css
@@ -1,0 +1,47 @@
+.imageWrapper {
+  padding: 24px;
+}
+
+.info {
+  padding: 24px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  font-family: 'Courier New', Courier, monospace;
+}
+
+.infoLine {
+  display: inline-block;
+}
+
+.bold {
+  font-weight: 600;
+}
+
+.marginBottom {
+  margin-block-end: 12px;
+}
+
+.focusClass {
+  height: 100%;
+  width: 100%;
+}
+
+.contentClass {
+  height: 100%;
+  width: 100%;
+  overflow-y: auto !important;
+  max-width: none !important;
+  max-height: none !important;
+  border: none !important;
+  border-radius: 0 !important;
+  background-color: transparent !important;
+  font-family: 'Courier New', Courier, monospace;
+}
+
+
+.flexColumn {
+  display: flex;
+  flex-direction: column;
+  margin-block-end: 24px;
+}
