+++ src/components/patient-record/patient-record.types.ts
@@ -1,0 +1,10 @@
+export type PatientRecordProps = {
+  name: string;
+  id: number;
+  birthDate: string;
+  birthPlace: string;
+  checkIn: string;
+  text: string;
+  withPic?: boolean;
+  withButton?: boolean;
+};
