+++ src/components/rotating-emblem/rotating-emblem.types.ts
@@ -1,0 +1,3 @@
+export type RotatingEmblemProps = {
+  txt: string;
+};
