+++ src/pages/index.tsx
@@ -1,87 +1,131 @@
 import React, { useRef } from "react";
 import cx from "classnames";
-// import { Parallax, ParallaxLayer } from "@react-spring/parallax";
-import styles from "../styles/About.module.css";
+import styles from "../styles/Home.module.css";
 import { Title } from "@/components/title";
-import { AdminProfile } from "@/components/admin-profile";
-import { Address, multicall, readContract } from "@wagmi/core";
-import { MOONPAGE_PROFILES_ADDRESS_DEV } from "@/constants";
-import ABI from "../abis/MoonpageProfiles.json";
-import { loop } from "@/utils/ loop";
-import { ADMIN_ADDRESSES } from "@/constants";
-import { Abi } from "abitype";
-import { Call } from "@/components/slider/slider.types";
-import { Profile } from "@/hooks/use-profile/use-profile.types";
 import { inter } from "@/utils/fonts";
 import Image from "next/image";
 import { Card } from "@/components/card";
+import { Button } from "@/components/button";
+import { RotatingEmblem } from "@/components/rotating-emblem/rotating-emblem";
+import Marquee from "react-fast-marquee";
+import { Folder } from "@/components/folder";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faStar } from "@fortawesome/free-solid-svg-icons";
 
-const call = {
-  address: MOONPAGE_PROFILES_ADDRESS_DEV as Address,
-  abi: ABI,
-  functionName: "profiles",
-};
+const StarRating = () => (
+  <div className={styles.starRating}>
+    <FontAwesomeIcon icon={faStar} size="xs" className={styles.star} />
+    <FontAwesomeIcon icon={faStar} size="xs" className={styles.star} />
+    <FontAwesomeIcon icon={faStar} size="xs" className={styles.star} />
+    <FontAwesomeIcon icon={faStar} size="xs" className={styles.star} />
+    <FontAwesomeIcon icon={faStar} size="xs" className={styles.star} />
+  </div>
+);
 
-let contracts: Call[] = [];
-loop(ADMIN_ADDRESSES.length, (i: number) => {
-  // @ts-ignore
-  contracts.push({ ...call, args: [ADMIN_ADDRESSES[i] as Address] });
-});
-
-export async function getServerSideProps(context: any) {
-  let data = await multicall({ contracts });
-  // @ts-ignore
-  data = data?.map((x, idx) => ({ ...x, address: ADMIN_ADDRESSES[idx] }));
-
-  return {
-    props: { profiles: data },
-  };
-}
-
-const Home = ({ profiles }: any) => {
+const Home = () => {
   return (
-    <div className={cx(styles.about, inter.variable)}>
-      <div className={cx(styles.first, styles.block, styles.withMarquee)}>
-        <div className={cx(styles.card, styles.logoWrapper)}>
-          <Image
-            src={"Logo.svg"}
-            height={200}
-            width={200}
-            alt={"TheRetreat Logo"}
-            priority
-          />
-          <span className={styles.text}>The one retreat you need.</span>
+    <div className={cx(styles.background, inter.className)}>
+      <div className={cx(styles.foreground, inter.className)}>
+        <div className={styles.section}>
+          <div className={styles.control}>
+            <div className={styles.redCircle}></div>
+            <div className={styles.titleWrapper}>
+              <Title
+                size={1}
+                className={cx(
+                  styles.title,
+                  styles.skewingAnimation,
+                  styles.asianTitle,
+                  styles.firstTitle
+                )}
+              >
+                退却
+              </Title>
+              <div className={styles.westernTitle}>
+                <Title size={1} className={cx(styles.title, styles.firstTitle)}>
+                  THE RETREAT
+                </Title>
+                <Title size={3} className={styles.subtitle}>
+                  The one retreat you need.
+                </Title>
+              </div>
+            </div>
+          </div>
         </div>
-      </div>
-      <div className={cx(styles.second, styles.block)}>
-        <div className={styles.card}>
-          <Title size={1} className={styles.aboutTitle}>
-            What is TheRetreat?
-          </Title>
-          <span className={styles.text}>
-            A high-end space sanatorial retreat program to upgrade your inner
-            self. 3 months in a space capsule rotating around planet earth.
-          </span>
-          <p className={styles.nextLaunchInfo}> Next Launch: 1st June.</p>
+        <div className={styles.section}>
+          <div className={styles.textInfo}>
+            <div className={styles.titlesWrapper}>
+              <Title size={1} className={styles.title}>
+                Self-discovery, healing and renewal.
+              </Title>
+              <Title size={3} className={styles.subtitle}>
+                "Because you must make yourself fantastic in order to do
+                fantastic things." – SadGhuru2001
+              </Title>
+            </div>
+            <Button className={styles.joinButton}>Yes, please</Button>
+          </div>
+          <div className={styles.statsInfo}>
+            <div className={styles.stat}>
+              <Title size={1} className={styles.statPercentage}>
+                93%
+              </Title>
+              <Title size={3} className={styles.subtitle}>
+                beings suffering mentally
+              </Title>
+            </div>
+            <div className={styles.stat}>
+              <Title size={1} className={styles.statPercentage}>
+                120
+              </Title>
+              <Title size={3} className={styles.subtitle}>
+                yearly patients
+              </Title>
+            </div>
+            <div className={styles.stat}>
+              <Title size={1} className={styles.statPercentage}>
+                125
+              </Title>
+              <Title size={3} className={styles.subtitle}>
+                yearly healed patients
+              </Title>
+            </div>
+          </div>
         </div>
-      </div>
-      <div className={cx(styles.third, styles.block)}>
-        <div className={styles.card}>
-          <Title size={1} className={cx(styles.aboutTitle, styles.vision)}>
+        <div className={styles.section}>
+          <div className={styles.orbitContainer}>
+            <div className={styles.innerOrbit}>
+              <div className={styles.innerOrbitCirlces}></div>
+            </div>
+            <div className={styles.outerOrbit}>
+              <div className={styles.outerOrbitCirlces}></div>
+            </div>
+          </div>
+          <div className={styles.titleWrapper}>
+            <Title size={1} className={styles.title}>
+              What is The Retreat?
+            </Title>
+            <br />
+            <Title size={3} className={styles.subtitle}>
+              A high-end space sanatorial retreat program to upgrade your inner
+              self. 3 months in a space capsule rotating around planet earth.
+              Next Launch: 1st June.
+            </Title>
+          </div>
+        </div>
+        <div className={cx(styles.section, styles.explanationSection)}>
+          <Title size={1} className={styles.title}>
             How does it work?
           </Title>
-          <p className={cx(styles.nextLaunchInfo, styles.centerAlign)}>
-            {" "}
-            Next Launch: 1st June.
-          </p>
-          <div className={styles.cardWrapper}>
+
+          <div className={styles.howitworks}>
             <Card>
               <Title size={3} className={styles.cardTitle}>
                 1.
               </Title>
               <span className={styles.cardText}>
-                Our independent algorithm handpicks the 20 participants for the
-                upcoming retreat 3 months in advance and notifies them.
+                We facilitate 4 retreats a year with a small group of selected
+                patients.
               </span>
             </Card>
             <Card>
@@ -89,10 +133,8 @@ const Home = ({ profiles }: any) => {
                 2.
               </Title>
               <span className={styles.cardText}>
-                We facilitate 4 retreats a year with a maximum capacity of 20
-                patients per retreat. Apply today by filling out this form,
-                because spots are in high demand! See the pricing list here or
-                join the lottery.
+                You can apply by upgrading to our platinum tier to secure a spot
+                in our lottery.
               </span>
             </Card>
             <Card>
@@ -100,9 +142,8 @@ const Home = ({ profiles }: any) => {
                 3.
               </Title>
               <span className={styles.cardText}>
-                Once you get picked, you will be notified and invited to our
-                screening, where you will set a personal goal for the retreat. A
-                healing curriculum will be generated for you.
+                Our independent algorithm handpicks the 20 participants for the
+                upcoming retreat.
               </span>
             </Card>
             <Card>
@@ -110,48 +151,122 @@ const Home = ({ profiles }: any) => {
                 4.
               </Title>
               <span className={styles.cardText}>
-                Off you go to space for three months, following the curriculum,
-                accompanied by our high profile scientist. The SoulScape space
-                capsule rotates around planet earth.
+                If you get picked, you will be invited to our screening, where
+                you will set a personal goal for the retreat. A healing
+                curriculum will be generated.
               </span>
             </Card>
+            <Card>
+              <Title size={3} className={styles.cardTitle}>
+                5.
+              </Title>
+              <span className={styles.cardText}>
+                Off you go to space for 3 months, following the curriculum,
+                accompanied by our high profile scientist.
+              </span>
+            </Card>
+            <RotatingEmblem txt="trusted*space*retreat*" />
           </div>
         </div>
-      </div>
-      <div className={cx(styles.fourth, styles.block)}>
-        <div className={styles.card}>
-          <Title size={1} className={cx(styles.aboutTitle, styles.vision)}>
-            Vision
-          </Title>
-          <span className={styles.text}>
-            This is Sadghuru2001, your AI CEO of TheRetreat. Our vision is to
-            weave a sanctuary of tranquility in space where one can untangle
-            from the cacophony of modern life and bask in the serenity of
-            self-discovery, connection, and renewal. Because you must make
-            yourself fantastic in order to do fantastic things. This is our firm
-            belief since 2023 and the reason we are market leader eversince.
-            Join our Space SoulScape Retreat program to heal and upgrade your
-            inner self.
-          </span>
+        <div className={styles.section}>
+          <div className={styles.titleWrapper}>
+            <Title size={1} className={styles.title}>
+              Our Vision
+            </Title>
+            <br />
+            <Title size={3} className={styles.subtitle}>
+              This is SadGhuru2001, your AI CEO of TheRetreat. Our vision is to
+              weave a sanctuary of tranquility in space where one can untangle
+              from the cacophony of modern life and bask in the serenity of
+              self-discovery, connection, and renewal. Because you must make
+              yourself fantastic in order to do fantastic things. This is our
+              firm belief since 2023 and the reason we are market leader
+              eversince. Join our Space SoulScape Retreat program to heal and
+              upgrade your inner self.
+            </Title>
+          </div>
+          <div className={styles.guruImageWrapper}>
+            <div className={styles.redFilter} />
+            <div className={styles.glitchLayers}>
+              <div className={styles.glitchLayer}></div>
+              <div className={styles.glitchLayer}></div>
+              <div className={styles.glitchLayer}></div>
+              <div className={styles.glitchLayer}></div>
+            </div>
+            <Image
+              height={300}
+              width={300}
+              src="/characters/guru.jpeg"
+              alt={`Placeholder Image`}
+              priority
+              className={styles.guruImage}
+            />
+          </div>
         </div>
-      </div>
-      <div className={cx(styles.fifth, styles.block)}>
-        <div className={styles.card}>
-          <Title size={1} className={cx(styles.aboutTitle, styles.vision)}>
-            Cut the BS, what is this, really?
+        <div className={cx(styles.section, styles.explanationSection)}>
+          <Title size={1} className={cx(styles.title, styles.reviewsTitle)}>
+            Patients Reviews
           </Title>
-          <p className={cx(styles.nextLaunchInfo, styles.centerAlign)}>
-            {" "}
-            NFT Launch: 1st June.
-          </p>
-          <div className={styles.cardWrapper}>
+          <Marquee>
+            <Folder className={styles.reviewFolder}>
+              <div className={styles.customerReview}>
+                <Title size={3}>Geno Feno</Title>
+                <br />
+                <span>
+                  Two great decisions I've made in my life. Doing TheRetreat and
+                  getting a divorce. I bought my new husband a platinum ticket.
+                </span>
+                <StarRating />
+              </div>
+            </Folder>
+            <Folder className={styles.reviewFolder}>
+              <div className={styles.customerReview}>
+                <Title size={3}>M. K.</Title>
+                <br />
+                <span>
+                  I was suffering from severe burnout. And this retreat helped
+                  me find my balance. SadGhuru was right: you must be fantastic
+                  in order to do fantastic things!
+                </span>
+                <StarRating />
+              </div>
+            </Folder>
+            <Folder className={styles.reviewFolder}>
+              <div className={styles.customerReview}>
+                <Title size={3}>Annie Hwang</Title>
+                <br />
+                <span>
+                  Speaking for a friend, since I personally do not have any
+                  mental issues. Highly recommended! No surprise this space
+                  retreat is market leader.
+                </span>
+                <StarRating />
+              </div>
+            </Folder>
+            <Folder className={styles.reviewFolder}>
+              <div className={styles.customerReview}>
+                <Title size={3}>0x9ad7...f1351</Title>
+                <br />
+                <span>
+                  Erstaunlich! Ich war multi-morbid und nun heiler als je zuvor.
+                </span>
+                <StarRating />
+              </div>
+            </Folder>
+          </Marquee>
+        </div>
+        <div className={cx(styles.section, styles.explanationSection)}>
+          <Title size={1} className={styles.title}>
+            Ok, what is this, really?
+          </Title>
+          <div className={styles.howitworks}>
             <Card>
               <Title size={3} className={styles.cardTitle}>
                 1.
               </Title>
               <span className={styles.cardText}>
-                We have set up the story frame for you and roughly drafted 10
-                characters.
+                We have set up this story frame for you and roughly drafted 10
+                characters. They are patients participating in TheRetreat.
               </span>
             </Card>
             <Card>
@@ -171,8 +286,8 @@ const Home = ({ profiles }: any) => {
               <span className={styles.cardText}>
                 Which character of the 10 it is, depends on the NFT you minted.
                 No traits and rarity scores, meaning no NFT or character is
-                rarer that the other. You can only write with an NFT once. If
-                you want to continue writing, you have to mint another NFT.
+                rarer than another. You can write with an NFT once. If you want
+                to continue writing, you have to mint another.
               </span>
             </Card>
             <Card>
@@ -180,31 +295,23 @@ const Home = ({ profiles }: any) => {
                 4.
               </Title>
               <span className={styles.cardText}>
-                The story is being written collectively, you contribute to the
-                piece of literature with the NFT.
+                The story grows through collective contribution. Your text
+                becomes part of literary NFT art.
               </span>
             </Card>
           </div>
         </div>
-      </div>
-      <div className={cx(styles.sixth, styles.block)}>
-        <div className={styles.card}>
-          <Title size={1} className={styles.aboutTitle}>
-            About us
-          </Title>
-          <span className={styles.text}>
-            We are Creative Writing students from the University of Hildesheim
-            experimenting with literary NFTs in collaboration with{" "}
-            <a
-              href="https://moonpage.io"
-              target="_blank"
-              rel="noreferrer"
-              className={styles.link}
-            >
-              Moonpage
-            </a>
-            .
-          </span>
+        <div className={styles.section}>
+          <div className={styles.titleWrapper}>
+            <Title size={1} className={styles.title}>
+              About us
+            </Title>
+            <br />
+            <Title size={3} className={styles.subtitle}>
+              We are Creative Writing students from the University of Hildesheim
+              experimenting with literary NFTs in collaboration with Moonpage.
+            </Title>
+          </div>
         </div>
       </div>
     </div>
